+++ pdm/builders/__init__.py
@@ -1,0 +1,9 @@
+from pdm.builders.egg_info import EnvEggInfoBuilder
+from pdm.builders.sdist import EnvSdistBuilder
+from pdm.builders.wheel import EnvWheelBuilder
+
+__all__ = (
+    EnvEggInfoBuilder.__name__,
+    EnvSdistBuilder.__name__,
+    EnvWheelBuilder.__name__,
+)
