+++ pdm/models/environment.py
@@ -17,7 +17,6 @@ from pip._vendor import packaging, pkg_resources
 from pdm import termui
 from pdm.models import pip_shims
 from pdm.models.auth import make_basic_auth
-from pdm.models.builders import EnvBuilder
 from pdm.models.in_process import (
     get_pep508_environment,
     get_python_abi_tag,
@@ -225,6 +224,8 @@ class Environment:
         :param allow_all: Allow building incompatible wheels.
         :returns: The full path of the built artifact.
         """
+        from pdm.builders import EnvEggInfoBuilder, EnvWheelBuilder
+
         build_dir = self._get_build_dir(ireq)
         wheel_cache = self.project.make_wheel_cache()
         supported_tags = pip_shims.get_supported(self.interpreter.for_tag())
@@ -282,30 +283,28 @@ class Environment:
                 return cache_entry.link.file_path
 
         # Otherwise, as all source is already prepared, build it.
-        with EnvBuilder(ireq.unpacked_source_directory, self) as builder:
-            if ireq.editable:
-                ret = builder.build_egg_info(build_dir)
-                ireq.metadata_directory = ret
+        if ireq.editable:
+            with EnvEggInfoBuilder(ireq.unpacked_source_directory, self) as builder:
+                ret = ireq.metadata_directory = builder.build(build_dir)
                 return ret
-            should_cache = False
-            if ireq.link.is_vcs:
-                vcs = pip_shims.VcsSupport()
-                vcs_backend = vcs.get_backend_for_scheme(ireq.link.scheme)
-                if vcs_backend.is_immutable_rev_checkout(
-                    ireq.link.url, ireq.source_dir
-                ):
-                    should_cache = True
-            else:
-                base, _ = ireq.link.splitext()
-                if _egg_info_re.search(base) is not None:
-                    # Determine whether the string looks like an egg_info.
-                    should_cache = True
-            output_dir = (
-                wheel_cache.get_path_for_link(ireq.link) if should_cache else build_dir
-            )
-            if not os.path.exists(output_dir):
-                os.makedirs(output_dir, exist_ok=True)
-            return builder.build_wheel(output_dir)
+        should_cache = False
+        if ireq.link.is_vcs:
+            vcs = pip_shims.VcsSupport()
+            vcs_backend = vcs.get_backend_for_scheme(ireq.link.scheme)
+            if vcs_backend.is_immutable_rev_checkout(ireq.link.url, ireq.source_dir):
+                should_cache = True
+        else:
+            base, _ = ireq.link.splitext()
+            if _egg_info_re.search(base) is not None:
+                # Determine whether the string looks like an egg_info.
+                should_cache = True
+        output_dir = (
+            wheel_cache.get_path_for_link(ireq.link) if should_cache else build_dir
+        )
+        if not os.path.exists(output_dir):
+            os.makedirs(output_dir, exist_ok=True)
+        with EnvWheelBuilder(ireq.unpacked_source_directory, self) as builder:
+            return builder.build(output_dir)
 
     def get_working_set(self) -> WorkingSet:
         """Get the working set based on local packages directory."""
